Save project changes : manager cand add + edit products  cart + order done   the home page is done.   to do:    socket     fix responsivity      add upload image       download a txt file of the recite

+++ server/models/CartItemsModel.js
@@ -9,7 +9,7 @@ const CartItem = sequelize.define('cart_Items', {
         primaryKey: true,
     },
     total_price: {
-        type: Sequelize.INTEGER(11),
+        type: Sequelize.DOUBLE,
         defaultValue: 0
     },
     qnt: {
